fix import rnnrollout
