+++ tatk/e2e/rnn_rollout/__init__.py
@@ -1,0 +1,1 @@
+from tatk.e2e.rnn_rollout.rnnrollout import RNNRolloutAgent
