+++ tatk/e2e/rnn_rollout/models/attn.py
@@ -11,7 +11,7 @@ import torch.nn.functional as F
 import torch.nn.init
 from torch.autograd import Variable
 
-from tatk.e2e.rnn_rullout.models.utils import *
+from tatk.e2e.rnn_rollout.models.utils import *
 
 
 class Attention(nn.Module):
