+++ tatk/e2e/rnn_rollout/models/latent_clustering_model.py
@@ -16,16 +16,16 @@ import torch.nn.init
 from torch.autograd import Variable
 import torch.nn.functional as F
 
-import tatk.e2e.rnn_rullout.data as data
-import tatk.e2e.rnn_rullout.utils
-from tatk.e2e.rnn_rullout.domain import get_domain
-from tatk.e2e.rnn_rullout.models.ctx_encoder import MlpContextEncoder
-from tatk.e2e.rnn_rullout.models.utils import *
+import tatk.e2e.rnn_rollout.data as data
+import tatk.e2e.rnn_rollout.utils
+from tatk.e2e.rnn_rollout.domain import get_domain
+from tatk.e2e.rnn_rollout.models.ctx_encoder import MlpContextEncoder
+from tatk.e2e.rnn_rollout.models.utils import *
 
 
 
-from tatk.e2e.rnn_rullout.engines.latent_clustering_engine import LatentClusteringEngine, LatentClusteringPredictionEngine
-from tatk.e2e.rnn_rullout.engines.latent_clustering_engine import LatentClusteringLanguageEngine, BaselineClusteringEngine
+from tatk.e2e.rnn_rollout.engines.latent_clustering_engine import LatentClusteringEngine, LatentClusteringPredictionEngine
+from tatk.e2e.rnn_rollout.engines.latent_clustering_engine import LatentClusteringLanguageEngine, BaselineClusteringEngine
 
 
 class SimpleSeparateSelectionModule(nn.Module):
