+++ tatk/e2e/rnn_rollout/models/rnn_model.py
@@ -16,10 +16,11 @@ import torch.nn.init
 from torch.autograd import Variable
 import torch.nn.functional as F
 
-import data
-from tatk.e2e.rnn_rullout.engines.rnn_engine import RnnEngine
-from tatk.e2e.rnn_rullout.utils import *
-from tatk.e2e.rnn_rullout.models.ctx_encoder import MlpContextEncoder
+import tatk.e2e.rnn_rollout.data as data
+from tatk.e2e.rnn_rollout.engines.rnn_engine import RnnEngine
+from tatk.e2e.rnn_rollout.domain import get_domain
+from tatk.e2e.rnn_rollout.models.utils import *
+from tatk.e2e.rnn_rollout.models.ctx_encoder import MlpContextEncoder
 
 
 class RnnModel(nn.Module):
