--- tatk/e2e/rnn_rullout/__init__.py
@@ -1,1 +1,0 @@
-from tatk.e2e.rnn_rullout.rnnrollout import RNNRolloutAgent
